feat: validation

+++ Citronix/Citronix/src/main/java/com/app/Citronix/Model/DTO/Request/FermeRequest.java
@@ -3,14 +3,14 @@ package com.app.Citronix.Model.DTO.Request;
 import java.time.LocalDate;
 
 import lombok.*;
+import javax.validation.constraints.*;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class FermeRequest {
-
-    private int id;
+    private Long id;
     private String nom;
     private String adress;
     private Double superficie;
